+++ src/App.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import "./App.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import TypingPracticePage from "@pages/TypingPractice";
-import LoginPage from "@pages/Login";
 import Modals from "@components/Modals";
 
 function App() {
@@ -11,7 +10,6 @@ function App() {
       <Router>
         <Routes>
           <Route path="/" element={<TypingPracticePage />} />
-          <Route path="/login" element={<LoginPage />} />
         </Routes>
       </Router>
       <Modals />