+++ src/components/Modals.tsx
@@ -1,9 +1,11 @@
 import LessonEditorModal from "./LessonEditor";
+import LoginModal from "./LoginModal";
 
 export default function Modals(): JSX.Element | null {
   return (
     <>
       <LessonEditorModal />
+      <LoginModal />
     </>
   );
 }
@@ -19,7 +21,7 @@ export interface ModalContentProps {
   title: string;
   full?: boolean;
   grow?: boolean;
-  buttons: React.ReactNode[];
+  buttons?: React.ReactNode[];
 }
 
 export function ModalContent(props: ModalContentProps): JSX.Element {
@@ -33,14 +35,16 @@ export function ModalContent(props: ModalContentProps): JSX.Element {
           <h1 className="text-2xl mb-5">{props.title}</h1>
           {props.children}
           {props.grow && <div className="grow"></div>}
-          <div className="mt-5 flex gap-5">
-            <div className="grow"></div>
-            {props.buttons.map((button, index) => (
-              <div key={index} className="">
-                {button}
-              </div>
-            ))}
-          </div>
+          {props.buttons && (
+            <div className="mt-5 flex gap-5">
+              <div className="grow"></div>
+              {props.buttons.map((button, index) => (
+                <div key={index} className="">
+                  {button}
+                </div>
+              ))}
+            </div>
+          )}
         </div>
       </div>
     </>