+++ src/components/basic/button.tsx
@@ -1,5 +1,6 @@
 interface ButtonProps {
   text: string;
+  className?: string;
   icon?: React.ReactNode;
   styling?: ButtonStyles;
   onClick: () => void;
@@ -13,36 +14,26 @@ type ButtonStyles =
   | "disabled";
 
 export default function Button(props: ButtonProps): JSX.Element {
-  let bg = "";
-  let hover = "";
-  let active = "";
+  let classNameStyling = "";
   switch (props.styling) {
     case "positive":
-      bg = "bg-green-600";
-      hover = "hover:bg-green-500";
-      active = "active:bg-green-700";
+      classNameStyling = "bg-green-600 hover:bg-green-500 active:bg-green-700";
       break;
     case "primary":
-      bg = "bg-blue-600";
-      hover = "hover:bg-blue-500";
-      active = "active:bg-blue-700";
+      classNameStyling = "bg-blue-600 hover:bg-blue-500 active:bg-blue-700";
       break;
     case "negative":
-      bg = "bg-red-600";
-      hover = "hover:bg-red-500";
-      active = "active:bg-red-700";
+      classNameStyling = "bg-red-600 hover:bg-red-500 active:bg-red-700";
       break;
     case "neutral":
     default:
-      bg = "bg-gray-700";
-      hover = "hover:bg-gray-600";
-      active = "active:bg-gray-800";
+      classNameStyling = "bg-gray-700 hover:bg-gray-600 active:bg-gray-800";
       break;
   }
 
   return (
     <button
-      className={`p-3 uppercase font-medium rounded-md min-w-36 ${bg} ${hover} ${active}`}
+      className={`p-3 flex items-center uppercase font-medium rounded-md min-w-36 ${classNameStyling} ${props.className}`}
       onClick={() => props.onClick()}
     >
       {props.icon && props.icon}