+++ src/pages/TypingPractice.tsx
@@ -7,11 +7,12 @@ import useWpmCounterStore from "@stores/wpmCounter";
 import useSettingsStore from "@stores/settings";
 import React, { useEffect } from "react";
 import { LessonMenu, LessonMenuClosed } from "@components/LessonMenu";
+import TopBar from "@components/TopBar";
 
 export default function TypingPracticePage(): JSX.Element {
   return (
-    <>
-      <LessonMenuClosed />
+    <div className="h-full w-full flex flex-col">
+      <TopBar menu={<LessonMenuClosed />} />
       <div className="h-full w-full flex">
         <LessonMenu />
         <div className="grow"></div>
@@ -22,7 +23,7 @@ export default function TypingPracticePage(): JSX.Element {
         </div>
         <div className="grow"></div>
       </div>
-    </>
+    </div>
   );
 }
 
@@ -184,15 +185,16 @@ function TypingField(): JSX.Element {
 
   function handleChange(event: React.ChangeEvent<HTMLInputElement>) {
     const value = event.target.value;
-    wpmCounter.updateCharTime(value.length - 1);
-
-    if (!startTime) {
-      startTimer();
-    }
 
     // Sometimes we want to ignore space if it's the first character and it's not part of the "word"
     if (value.length === 1 && value === " " && correctInput[0] !== " ") {
       return;
+    }
+
+    wpmCounter.updateCharTime(value.length - 1);
+
+    if (!startTime) {
+      startTimer();
     }
 
     if (value === correctInput) {