+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### ➕ Added
+
+- Top menu with profile and login/logout menu
+- Logout functionality
+
+### 🔄 Changed
+
+- Login is done through a modal instead of /login
+- Increase the WPM multiplier from 0.5 to 3 when calculating next words to practice
+
+### 🐛 Fixed
+
+- Pressing space at the start does no longer clear the current WPM
+
 ## 0.1.0 - 2024-07-03
 
 ### ➕ Added